extract `with_history()` to separated module (#4)
ref #4
